Updated deployment setup

+++ app/schemas.py
@@ -1,18 +1,18 @@
 from pydantic import BaseModel
 
 class RentalInput(BaseModel):
-    city: str  # Included for input schema, not used in model
+    city: str
     area: str
     location: str
     zone: str
     property_type: str
-    size_in_sqft: int
+    size_in_sqft: float
     bedrooms: int
     bathrooms: int
     balcony: int
     furnishing_status: str
     Number_of_amenities: int
-    security_deposite: int
+    security_deposite: float
     property_age: int
     brokerage: str
     floor_no: int
@@ -20,4 +20,3 @@ class RentalInput(BaseModel):
     nearby_facilities: str
     type_of_society: str
     road_connectivity: str
-
